refactor: simplify and unify module names
