+++ packages/react/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./ReactPlainBottomSheet";
+export * from "./BottomSheet";
